requirements.txt
# Core Flask Dependencies
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0

# Authentication & Security
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
bcrypt==5.0.0

# Database
PyMongo==4.15.3
dnspython==2.8.0

# Environment & Configuration
python-dotenv==1.1.1

# HTTP Requests (for API integration)
requests==2.31.0

# Additional API dependencies
spotipy==2.23.0  # Spotify Web API wrapper (optional)
google-api-python-client==2.110.0  # YouTube API client (optional)

# Development Dependencies (optional)
# Flask-DebugToolbar==0.13.1
# pytest==7.4.0
# pytest-flask==1.3.0
